Refactor state interval handling by moving StateInterval and related functions to a new state_intervals.py file. Updated import statements across the area occupancy component to reflect this change, enhancing code organization and maintainability.
