+++ custom_components/area_occupancy/utils.py
@@ -2,17 +2,9 @@
 
 from __future__ import annotations
 
-from collections.abc import Sequence
 from datetime import datetime, timedelta
-import logging
-from typing import TYPE_CHECKING, Any, TypedDict
+from typing import TYPE_CHECKING
 
-from sqlalchemy.exc import SQLAlchemyError
-
-from homeassistant.components.recorder.history import get_significant_states
-from homeassistant.core import HomeAssistant, State
-from homeassistant.exceptions import HomeAssistantError
-from homeassistant.helpers.recorder import get_instance
 from homeassistant.util import dt as dt_util
 
 from .const import (
@@ -25,102 +17,6 @@ from .const import (
 
 if TYPE_CHECKING:
     from .data.entity import Entity
-
-_LOGGER = logging.getLogger(__name__)
-
-# Interval filtering thresholds to exclude anomalous data
-# Exclude intervals shorter than 10 seconds (false triggers)
-MIN_INTERVAL_SECONDS = 10
-# Exclude intervals longer than 13 hours (stuck sensors)
-MAX_INTERVAL_SECONDS = 13 * 3600
-# States to exclude from intervals
-INVALID_STATES = {"unknown", "unavailable", None, "", "NaN"}
-
-
-class Interval(TypedDict):
-    """Time interval with state information."""
-
-    start: datetime
-    end: datetime
-
-
-class StateInterval(Interval):
-    """Time interval with time information."""
-
-    state: str
-    entity_id: str
-
-
-class PriorInterval:
-    """Time interval with prior information."""
-
-    start_hour: int
-    start_minute: int
-    end_hour: int
-    end_minute: int
-    prior: float | None
-
-    def __init__(
-        self,
-        start_hour: int,
-        start_minute: int,
-        end_hour: int,
-        end_minute: int,
-        prior: float | None,
-    ) -> None:
-        """Initialize prior interval."""
-        self.start_hour = start_hour
-        self.start_minute = start_minute
-        self.end_hour = end_hour
-        self.end_minute = end_minute
-        self.prior = prior
-
-    def to_interval(self) -> Interval:
-        """Convert to interval."""
-        return Interval(
-            start=datetime(
-                datetime.now().year,
-                datetime.now().month,
-                datetime.now().day,
-                self.start_hour,
-                self.start_minute,
-            ),
-            end=datetime(
-                datetime.now().year,
-                datetime.now().month,
-                datetime.now().day,
-                self.end_hour,
-                self.end_minute,
-            ),
-        )
-
-
-class LikelihoodInterval(Interval):
-    """Time interval with likelihood information."""
-
-    prob_given_true: float
-    prob_given_false: float
-
-
-def is_valid_state(state: Any) -> bool:
-    """Check if a state is valid."""
-    return state not in INVALID_STATES
-
-
-async def init_times_of_day(hass: HomeAssistant) -> list[PriorInterval]:
-    """Initialize times of day for a given entity as 24 hourly intervals."""
-    return [
-        PriorInterval(
-            start_hour=hour,
-            start_minute=minute,
-            end_hour=hour if minute == 0 else (hour + 1) % 24,
-            end_minute=29 if minute == 0 else 59,
-            prior=None,
-        )
-        for hour in range(24)
-        for minute in (0, 30)
-    ]
-
 
 # ──────────────────────────────────── Time-Based Prior Utilities ──────────────────────────────────
 
@@ -248,215 +144,15 @@ def get_all_time_slots() -> list[tuple[int, int]]:
     return [(day, slot) for day in range(7) for slot in range(48)]
 
 
-# ──────────────────────────────────── History Utilities ──────────────────────────────────
-async def get_intervals_from_recorder(
-    hass: HomeAssistant, entity_id: str, start_time: datetime, end_time: datetime
-) -> list[StateInterval]:
-    """Get intervals from HA recorder."""
-    _LOGGER.debug(
-        "Getting intervals from recorder for %s from %s to %s",
-        entity_id,
-        start_time,
-        end_time,
-    )
-
-    states = await get_states_from_recorder(hass, entity_id, start_time, end_time)
-
-    _LOGGER.debug(
-        "Got %d states from recorder for %s", len(states) if states else 0, entity_id
-    )
-
-    if not states:
-        _LOGGER.debug("No states found in recorder for %s", entity_id)
-        return []
-
-    # Filter to only State objects and exclude unavailable/unknown states
-    state_objects: list[State] = [
-        s for s in states if isinstance(s, State) and is_valid_state(s.state)
-    ]
-
-    _LOGGER.debug(
-        "Filtered to %d valid State objects for %s", len(state_objects), entity_id
-    )
-
-    intervals = await states_to_intervals(state_objects, start_time, end_time)
-
-    _LOGGER.debug("Converted to %d intervals for %s", len(intervals), entity_id)
-
-    return filter_intervals(intervals)
-
-
-async def get_states_from_recorder(
-    hass: HomeAssistant, entity_id: str, start_time: datetime, end_time: datetime
-) -> list[State | dict[str, Any]] | None:
-    """Fetch states history from recorder.
-
-    Args:
-        hass: Home Assistant instance
-        entity_id: Entity ID to fetch history for
-        start_time: Start time window
-        end_time: End time window
-
-    Returns:
-        List of states or minimal state dicts if successful, None if error occurred
-
-    Raises:
-        HomeAssistantError: If recorder access fails
-        SQLAlchemyError: If database query fails
-
-    """
-    _LOGGER.debug("Fetching states: %s [%s -> %s]", entity_id, start_time, end_time)
-
-    # Check if recorder is available
-    recorder = get_instance(hass)
-    if recorder is None:
-        _LOGGER.debug("Recorder not available for %s", entity_id)
-        return None
-
-    try:
-        _LOGGER.debug("Executing recorder query for %s", entity_id)
-        states = await recorder.async_add_executor_job(
-            lambda: get_significant_states(
-                hass,
-                start_time,
-                end_time,
-                [entity_id],
-                minimal_response=False,  # Must be false to include last_changed attribute
-            )
-        )
-
-        _LOGGER.debug("Recorder query completed for %s, processing results", entity_id)
-        entity_states = states.get(entity_id) if states else None
-
-        if entity_states:
-            _LOGGER.debug("Found %d states for %s", len(entity_states), entity_id)
-        else:
-            _LOGGER.debug("No states found for %s in recorder query result", entity_id)
-
-    except (HomeAssistantError, SQLAlchemyError, TimeoutError) as err:
-        _LOGGER.error("Error getting states for %s: %s", entity_id, err)
-        # Re-raise the exception as documented, let the caller handle fallback
-        raise
-
-    else:
-        return entity_states
-
-
-async def states_to_intervals(
-    states: Sequence[State], start: datetime, end: datetime
-) -> list[StateInterval]:
-    """Convert state history to time intervals.
-
-    Args:
-        states: List of State objects
-        start: Start time for analysis
-        end: End time for analysis
-
-    Returns:
-        List of StateInterval objects
-
-    """
-    intervals: list[StateInterval] = []
-    if not states:
-        return intervals
-
-    # Filter out any invalid states that might have slipped through
-    valid_states = [state for state in states if is_valid_state(state.state)]
-
-    if not valid_states:
-        _LOGGER.debug("No valid states after filtering unavailable/unknown")
-        return intervals
-
-    # Sort states chronologically
-    sorted_states = sorted(valid_states, key=lambda x: x.last_changed)
-
-    # Determine the state that was active at the start time
-    current_state = sorted_states[0].state
-    for state in sorted_states:
-        if state.last_changed <= start:
-            current_state = state.state
-        else:
-            break
-
-    current_time = start
-
-    # Build intervals between state changes
-    for state in sorted_states:
-        if state.last_changed <= start:
-            continue
-        if state.last_changed > end:
-            break
-
-        # Only create intervals for valid states
-        if is_valid_state(current_state):
-            intervals.append(
-                StateInterval(
-                    start=current_time,
-                    end=state.last_changed,
-                    state=current_state,
-                    entity_id=state.entity_id,
-                )
-            )
-        current_state = state.state
-        current_time = state.last_changed
-
-    # Final interval until end (only if state is valid)
-    if is_valid_state(current_state):
-        intervals.append(
-            StateInterval(
-                start=current_time,
-                end=end,
-                state=current_state,
-                entity_id=state.entity_id,
-            )
-        )
-
-    return intervals
-
-
-def filter_intervals(
-    intervals: list[StateInterval],
-) -> list[StateInterval]:
-    """Filter intervals to remove anomalous data and invalid states.
-
-    - For binary sensors ("on" state): filter out intervals that are too short or too long.
-    - For numeric sensors and other valid states: only remove intervals with invalid states, do not filter by duration.
-    """
-    # Remove intervals with invalid states
-    valid_intervals = [
-        interval for interval in intervals if is_valid_state(interval["state"])
-    ]
-
-    filtered_intervals = []
-    for interval in valid_intervals:
-        state = interval["state"]
-        # Only apply duration filtering to binary "on" state
-        if state == "on":
-            duration_seconds = (interval["end"] - interval["start"]).total_seconds()
-            if MIN_INTERVAL_SECONDS <= duration_seconds <= MAX_INTERVAL_SECONDS:
-                filtered_intervals.append(interval)
-            # else: skip this interval (too short/long)
-        else:
-            # For all other states (including numerics), keep as long as state is valid
-            filtered_intervals.append(interval)
-
-    return filtered_intervals
-
-
 # ───────────────────────────────────────── Validation ────────────────────────
 def validate_prob(value: complex) -> float:
     """Validate probability value, handling complex numbers."""
     # Handle complex numbers by taking the real part
     if isinstance(value, complex):
-        _LOGGER.warning(
-            "Complex number detected in probability calculation: %s, using real part",
-            value,
-        )
         value = value.real
 
     # Ensure it's a valid float
     if not isinstance(value, (int, float)) or not (-1e10 < value < 1e10):
-        _LOGGER.warning("Invalid probability value: %s, using default", value)
         return 0.5
 
     return max(0.001, min(float(value), 1.0))
